% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linmod.R
\name{linmod.default}
\alias{linmod.default}
\title{Estymacja MNK}
\usage{
\method{linmod}{default}(x, y, ...)
}
\arguments{
\item{x}{matrix}

\item{y}{numeric}

\item{...}{additional arguments to be passed in the future update of this package}
}
\value{
Funkcja zwaraca w wyniku parametry MNK
}
\description{
Estymacja MNK
}
\author{
Karolina Augustyniak, Aleksandra Chwalek
}
