% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linmod.R
\name{linmodEst}
\alias{linmodEst}
\title{Estymacja MNK}
\usage{
linmodEst(x, y)
}
\arguments{
\item{x}{matrix}

\item{y}{matrix}
}
\value{
Funkcja zwaraca w wyniku parametry MNK
}
\description{
Estymacja MNK
}
\author{
Karolina Augustyniak, Aleksandra Chwalek
}
