% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linmod.R
\name{print.linmod}
\alias{print.linmod}
\title{Wydruk MNK}
\usage{
\method{print}{linmod}(x, ...)
}
\arguments{
\item{x}{matrix}

\item{...}{additional arguments to be passed in the future update of this package}
}
\value{
Funkcja zwaraca w wyniku parametry MNK
}
\description{
Estymacja MNK
}
\author{
Karolina Augustyniak, Aleksandra Chwalek
}
