% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linmod.R
\name{summary.linmod}
\alias{summary.linmod}
\title{Podsumowanie MNK}
\usage{
\method{summary}{linmod}(object, ...)
}
\arguments{
\item{object}{matrix}

\item{...}{additional arguments to be passed in the future update of this package}
}
\value{
Funkcja zwaraca w wyniku podsumowanie MNK
}
\description{
Estymacja MNK
}
\author{
Karolina Augustyniak, Aleksandra Chwalek
}
